Added a .setdp directive to ca65's Assembly language.
Constant address operands that are within the 256 bytes at the address given by .setdp will be changed into direct-page operands.  That feature can be used when assembling code for the WDC65816 processor.  ca65 uses it automatically when assembling for the HuC6280.
When ca65 assembles code for the 65816, the directive will cause direct-page operands to become absolute addressing operands.